third_party/blink/web_tests/fast/css/shorthand-mismatched-list-crash-expected.txt

Test for WebKit bug 31559: Crash with mismatched lists and shorthands.

PASS para.style.webkitTransition is "width 1s, left 1s, 1s"
PASS para.style.webkitTransition is "width 1s, left 1s, top"
PASS successfullyParsed is true

TEST COMPLETE
